+++ allennlp/models/model.py
@@ -101,11 +101,24 @@ class Model(torch.nn.Module, Registrable):
                 raise RuntimeError("The regularizer cannot be a non-zero float.")
         return regularization_penalty
 
-    def get_parameters_for_histogram_tensorboard_logging(self) -> List[str]:
+    def get_parameters_for_histogram_logging(self) -> List[str]:
         """
         Returns the name of model parameters used for logging histograms to tensorboard.
         """
         return [name for name, _ in self.named_parameters()]
+
+    def get_parameters_for_histogram_tensorboard_logging(self) -> List[str]:
+        """
+        Returns the name of model parameters used for logging histograms to tensorboard.
+        """
+        import warnings
+
+        warnings.warn(
+            "'Model.get_parameters_for_histogram_tensorboard_logging' is deprecated, please use "
+            "'Model.get_parameters_for_histogram_logging' instead.",
+            DeprecationWarning,
+        )
+        return self.get_parameters_for_histogram_logging()
 
     def forward(self, *inputs) -> Dict[str, torch.Tensor]:
         """