+++ allennlp/modules/token_embedders/pretrained_transformer_embedder.py
@@ -38,7 +38,13 @@ class PretrainedTransformerEmbedder(TokenEmbedder):
         as embedders such as BERT. However, other models consist of encoder and decoder, in which case we just
         want to use the encoder.
     train_parameters: `bool`, optional (default = `True`)
-        If this is `True`, the transformer weights get updated during training.
+        If this is `True`, the transformer weights get updated during training. If this is `False`, the
+        transformer weights are not updated during training.
+    eval_mode: `bool`, optional (default = `False`)
+        If this is `True`, the model is always set to evaluation mode (e.g., the dropout is disabled and the
+        batch normalization layer statistics are not updated). If this is `False`, such dropout and batch
+        normalization layers are only set to evaluation mode when when the model is evaluating on development
+        or test data.
     last_layer_only: `bool`, optional (default = `True`)
         When `True` (the default), only the final layer of the pretrained transformer is taken
         for the embeddings. But if set to `False`, a scalar mix of all of the layers
@@ -64,6 +70,7 @@ class PretrainedTransformerEmbedder(TokenEmbedder):
         max_length: int = None,
         sub_module: str = None,
         train_parameters: bool = True,
+        eval_mode: bool = False,
         last_layer_only: bool = True,
         override_weights_file: Optional[str] = None,
         override_weights_strip_prefix: Optional[str] = None,
@@ -121,9 +128,24 @@ class PretrainedTransformerEmbedder(TokenEmbedder):
         self._num_added_end_tokens = len(tokenizer.single_sequence_end_tokens)
         self._num_added_tokens = self._num_added_start_tokens + self._num_added_end_tokens
 
+        self.train_parameters = train_parameters
         if not train_parameters:
             for param in self.transformer_model.parameters():
                 param.requires_grad = False
+
+        self.eval_mode = eval_mode
+        if eval_mode:
+            self.transformer_model.eval()
+
+    @overrides
+    def train(self, mode: bool = True):
+        self.training = mode
+        for name, module in self.named_children():
+            if self.eval_mode and name == "transformer_model":
+                module.eval()
+            else:
+                module.train(mode)
+        return self
 
     @overrides
     def get_output_dim(self):