+++ allennlp/predictors/sentence_tagger.py
@@ -26,7 +26,7 @@ class SentenceTaggerPredictor(Predictor):
         self, model: Model, dataset_reader: DatasetReader, language: str = "en_core_web_sm"
     ) -> None:
         super().__init__(model, dataset_reader)
-        self._tokenizer = SpacyTokenizer(language=language, pos_tags=True)
+        self._tokenizer = SpacyTokenizer(language=language)
 
     def predict(self, sentence: str) -> JsonDict:
         return self.predict_json({"sentence": sentence})