+++ tests/training/trainer_test.py
@@ -24,12 +24,15 @@ from allennlp.models.simple_tagger import SimpleTagger
 from allennlp.training import (
     GradientDescentTrainer,
     Checkpointer,
+)
+from allennlp.training.callbacks import (
     TrainerCallback,
     TrackEpochCallback,
     TensorBoardCallback,
-    SanityCheckCallback,
+    SanityChecksCallback,
     ConsoleLoggerCallback,
 )
+from allennlp.training.callbacks.sanity_checks import SanityCheckError
 from allennlp.training.learning_rate_schedulers import CosineWithRestarts
 from allennlp.training.learning_rate_schedulers import ExponentialLearningRateScheduler
 from allennlp.training.momentum_schedulers import MomentumScheduler
@@ -100,7 +103,7 @@ class TrainerTestBase(AllenNlpTestCase):
     def setup_method(self):
         super().setup_method()
         self.data_path = str(self.FIXTURES_ROOT / "data" / "sequence_tagging.tsv")
-        self.reader = SequenceTaggingDatasetReader()
+        self.reader = SequenceTaggingDatasetReader(max_instances=4)
         self.data_loader = MultiProcessDataLoader(self.reader, self.data_path, batch_size=2)
         self.data_loader_lazy = MultiProcessDataLoader(
             self.reader, self.data_path, batch_size=2, max_instances_in_memory=10
@@ -701,9 +704,9 @@ class TestTrainer(TrainerTestBase):
             num_epochs=3,
             serialization_dir=self.TEST_DIR,
             callbacks=[
-                TensorBoardCallback.from_params(
-                    Params({"tensorboard_writer": {"distribution_interval": 2}}),
+                TensorBoardCallback(
                     serialization_dir=self.TEST_DIR,
+                    distribution_interval=2,
                 )
             ],
         )
@@ -801,16 +804,10 @@ class TestTrainer(TrainerTestBase):
             num_epochs=2,
             serialization_dir=self.TEST_DIR,
             callbacks=[
-                TensorBoardCallback.from_params(
-                    Params(
-                        {
-                            "tensorboard_writer": {
-                                "summary_interval": 2,
-                                "should_log_learning_rate": True,
-                            }
-                        }
-                    ),
+                TensorBoardCallback(
                     serialization_dir=self.TEST_DIR,
+                    summary_interval=2,
+                    should_log_learning_rate=True,
                 )
             ],
         )
@@ -818,7 +815,6 @@ class TestTrainer(TrainerTestBase):
         trainer.train()
 
     def test_sanity_check_callback(self):
-
         model_with_bias = FakeModelForTestingNormalizationBiasVerification(use_bias=True)
         inst = Instance({"x": TensorField(torch.rand(3, 1, 4))})
         data_loader = SimpleDataLoader([inst, inst], 2)
@@ -828,9 +824,9 @@ class TestTrainer(TrainerTestBase):
             data_loader,
             num_epochs=1,
             serialization_dir=self.TEST_DIR,
-            callbacks=[SanityCheckCallback(serialization_dir=self.TEST_DIR)],
+            callbacks=[SanityChecksCallback(serialization_dir=self.TEST_DIR)],
         )
-        with pytest.raises(AssertionError):
+        with pytest.raises(SanityCheckError):
             trainer.train()
 
     def test_sanity_check_default(self):
@@ -843,7 +839,7 @@ class TestTrainer(TrainerTestBase):
             data_loader=data_loader,
             num_epochs=1,
         )
-        with pytest.raises(AssertionError):
+        with pytest.raises(SanityCheckError):
             trainer.train()
 
         trainer = GradientDescentTrainer.from_partial_objects(
@@ -851,7 +847,7 @@ class TestTrainer(TrainerTestBase):
             serialization_dir=self.TEST_DIR,
             data_loader=data_loader,
             num_epochs=1,
-            run_sanity_check=False,
+            run_sanity_checks=False,
         )
 
         # Check is not run, so no failure.
@@ -1070,7 +1066,7 @@ class TestTrainer(TrainerTestBase):
             def on_batch(
                 self,
                 trainer: "GradientDescentTrainer",
-                batch_inputs: List[List[TensorDict]],
+                batch_inputs: List[TensorDict],
                 batch_outputs: List[Dict[str, Any]],
                 batch_metrics: Dict[str, Any],
                 epoch: int,
@@ -1149,7 +1145,7 @@ class TestTrainer(TrainerTestBase):
             def on_batch(
                 self,
                 trainer: "GradientDescentTrainer",
-                batch_inputs: List[List[TensorDict]],
+                batch_inputs: List[TensorDict],
                 batch_outputs: List[Dict[str, Any]],
                 batch_metrics: Dict[str, Any],
                 epoch: int,
@@ -1195,23 +1191,15 @@ class TestTrainer(TrainerTestBase):
             num_epochs=2,
             serialization_dir=self.TEST_DIR,
             callbacks=[
-                TensorBoardCallback.from_params(
-                    Params(
-                        {
-                            "tensorboard_writer": {
-                                "distribution_interval": 2,
-                                "should_log_inputs": True,
-                            }
-                        }
-                    ),
+                TensorBoardCallback(
                     serialization_dir=self.TEST_DIR,
+                    distribution_interval=2,
                 )
             ],
         )
         trainer.train()
 
     def test_console_log_callback(self):
-
         total_instances = 1000
         batch_size = 25